worker: extract /api/hebraize into hebraize-route.ts (first index.ts slice)
Step 3 of the cleanup: begin splitting the 11.7k-line, 101-route worker/index.ts
into domain modules, following the established registerXRoutes(app) pattern
(commentary.ts). First slice is the self-contained one: POST /api/hebraize moves
verbatim into hebraize-route.ts along with HEBRAIZE_LLM_SYSTEM_PROMPT and the two
pure helpers (splitOuterWhitespace / sanitizeHebraizeOutput), which depend only on
imports (readJsonBody, runLLM, keyForHebraize, stripEchoParens). index.ts now just
calls registerHebraizeRoutes(app); the two now-dead imports it left behind are
removed.

(NB: the originally-floated pasuk/yerushalmi slice turned out coupled to shared
local helpers — cleanVerseText/getGemaraSlice — so it waits until those are pulled
into neutral modules; hebraize had no such coupling.)

The two hebraize unit tests now import from hebraize-route instead of index.
Behavior-only: -119 lines from index.ts, route logic byte-identical.
